[internal] Rename `Lessons`directory to `Lesson`to avoid cache issues
